backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
openai
apify-client
pandas
python-dotenv
google-generativeai
anthropic
tavily-python
